+++ data/key_annotation.tcl
@@ -6,3 +6,7 @@ set annotation [create_charts_key_annotation -plot $plot]
 add_charts_key_item -annotation $annotation -text Red   -color red
 add_charts_key_item -annotation $annotation -text Green -color green
 add_charts_key_item -annotation $annotation -text Blue  -color blue
+
+add_charts_key_item -annotation $annotation -text Cross  -symbol cross  -color red
+add_charts_key_item -annotation $annotation -text Circle -symbol circle -color green
+add_charts_key_item -annotation $annotation -text Box    -symbol box    -color blue